feat(calcresult): ✨Create CalcResult component
