+++ src/App.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import CalculationForm from './modules/CalculationForm';
+import CalcResult from './modules/CalcResult';
 import Header from './modules/Header';
 import { Base } from './styles/Base.js';
 
@@ -8,6 +9,7 @@ function App() {
     <Base>
       <Header />
       <CalculationForm />
+      <CalcResult />
     </Base>
   );
 }
