+++ src/modules/CalcResult.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const CalcResult = () => {
+  return (
+    <div>
+      <h2>Seu salário líquido será</h2>
+      <div>
+        <span>R$</span>
+        <span>0,00</span>
+      </div>
+    </div>
+  );
+};
+
+export default CalcResult;
