+++ src/modules/fields/components/ValueInput.jsx
@@ -1,31 +1,11 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { RealInput, RealSymbol } from '../../../styles/CalculationForm';
-
-const removeLetters = (string) => string.replace(/\D/g, '');
-
-/**
- * Add a cent separator to a common value
- * @param {string} value 234530
- * @returns {string} 2345.30
- */
-const addCentsSeparator = (value) => {
-  const centsIndex = value.length - 2;
-  const cents = value.substring(centsIndex);
-  const reais = value.substring(0, centsIndex);
-  return `${reais}.${cents}`;
-};
-
-/**
- * Converts a common value with separation of cents to Real
- * @param {string} value 2345.30
- * @returns {string} 2.345,30
- */
-const convertToReal = (value) => {
-  return Intl.NumberFormat('id', {
-    minimumFractionDigits: 2,
-  }).format(value);
-};
+import {
+  removeLetters,
+  addCentsSeparator,
+  convertToReal,
+} from '../../../utils';
 
 const invalidValues = ['', '0', '00', '000', '0000'];
